separati tempi minimi di accensione e di spegnimento pompa camino

+++ arduino/riscaldamento/progetto/sketches/_20170117_Override_Termostato_Ambiente/State.h
@@ -58,7 +58,7 @@ class State {
   unsigned long outOverrideTermoAmbienteAccu_On;
   unsigned long outOverrideTermoAmbienteAccu_Off;
   
-  long  ainTempCaminoValue;
+  unsigned int  ainTempCaminoValue;
   float ainTempCaminoValueCentigradi;
 };
